Extracted book form logic from add-book-component
