--- src/app/home/pages/add-book/add-book.component.scss
@@ -1,44 +1,0 @@
-
-.star {
-  color: gold;
-}
-
-#condition-label {
-  padding: 10px;
-  margin: 10px;
-}
-
-/*
-mat-card {
-  margin: 1% 25%;
-  width: 50%;
-  height: 98%;
-  border-radius: 10px;
-}
-
-mat-form-field {
-  width: 80%;
-  margin: auto 10%;
-}
-
-mat-card-title {
-  font-weight: bold;
-}
-
-button {
-  width: 20%;
-  margin: 0 40%;
-}
-
-.condition-container {
-  width: 50%;
-  margin: auto 25%;
-}
-
-.condition-container-label {
-  font: 400 14px/20px Roboto, sans-serif;
-  letter-spacing: .0178571429em;
-  color: rgba(0, 0, 0, 0.6);
-  margin: 20%;
-}
-*/
