+++ src/app/home/pages/inventory/inventory.component.html
@@ -1,1 +1,1 @@
-<p>inventory works!</p>
+<book-form [bookId]=1 pageName="Edit Book"/> <!--todo: pass bookId properly-->
